requirements.txt
clickhouse-sqlalchemy==0.3.0
pandas==1.5.3
python-dotenv==1.0.0
SQLAlchemy==2.0.27
textblob==0.18.0
tqdm==4.65.0
